test(book)! : test code 작성

+++ src/main/java/sesac/bookmanager/book/repository/BookRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import sesac.bookmanager.book.domain.Book;
 
 @Repository
-public interface BookRepository extends CrudRepository<Book, Integer>, BookQueryRepository {
+public interface BookRepository extends CrudRepository<Book, Long>, BookQueryRepository {
 }
